requirements.txt
pandas
numpy
matplotlib
scikit-learn
nltk
seaborn
textblob
tweepy
wordcloud
plotly
streamlit
praw
python-dotenv
vaderSentiment
joblib
openai
google-generativeai
openpyxl
xlrd<2.0.0
pyarrow
tables
lxml